完成考试提交业务，完成试卷回顾页面业务

+++ src/main/java/com/smileshark/entity/score/Score.java
@@ -16,5 +16,5 @@ public class Score {
     private String examPaperAllocationId;
     private ExamPaperAllocation examPaperAllocation;
     private Double score;
-    private Integer finishTime;
+    private Long finishTime;
 }
